Report Get iD changes

+++ HIMS.Data/Master/I_Hospital.cs
@@ -8,6 +8,6 @@ namespace HIMS.Data.Master
 {
     public interface I_Hospital
     {
-        HospitalMaster GetHospitalById(int Id);
+        HospitalMaster GetHospitalById(long Id);
     }
 }
